Move unknown cause triggers to new module
