+++ src/parser/core/modules/BrokenLog.tsx
@@ -7,8 +7,6 @@ import {getReportPatch} from 'data/PATCHES'
 import {ReactNode} from 'react'
 import {Table} from 'semantic-ui-react'
 import {Analyser, DisplayMode} from '../Analyser'
-import {dependency} from '../Injectable'
-import {Data} from './Data'
 import {DISPLAY_ORDER} from './DISPLAY_ORDER'
 
 interface Trigger {
@@ -21,8 +19,6 @@ export class BrokenLog extends Analyser {
 	static override title = t('core.broken-log.title')`Broken Log`
 	static override displayOrder = DISPLAY_ORDER.BROKEN_LOG
 	static override displayMode = DisplayMode.RAW
-
-	@dependency private data!: Data
 
 	private triggers = new Map<string, Trigger>()
 
@@ -69,24 +65,6 @@ export class BrokenLog extends Analyser {
 			source: constructor,
 			reason,
 		})
-	}
-
-	override initialise() {
-		const unknownAction = this.data.actions.UNKNOWN.id
-		this.addEventHook({cause: {type: 'action', action: unknownAction}}, this.triggerUnknownCause)
-		this.addEventHook({action: unknownAction}, this.triggerUnknownCause)
-
-		const unknownStatus = this.data.statuses.UNKNOWN.id
-		this.addEventHook({cause: {type: 'status', status: unknownStatus}}, this.triggerUnknownCause)
-		this.addEventHook({status: unknownStatus}, this.triggerUnknownCause)
-	}
-
-	private triggerUnknownCause() {
-		this.trigger(this, 'unknown action', (
-			<Trans id="core.broken-log.trigger.unknown-action">
-				One or more actions were recorded incorrectly, and could not be parsed.
-			</Trans>
-		))
 	}
 
 	override output() {
