+++ src/parser/core/modules/InvalidEvent.tsx
@@ -4,14 +4,25 @@ import {Event, PREPULL_EVENT_WINDOW} from "event"
 import {Analyser} from "../Analyser"
 import {dependency} from "../Injectable"
 import {BrokenLog} from "./BrokenLog"
+import {Data} from "./Data"
 
 export class InvalidEvent extends Analyser {
 	static override handle = 'invalidEvent'
 	static override title = t('core.invalid-event.title')`Invalid Event`
 
 	@dependency private brokenLog!: BrokenLog
+	@dependency private data!: Data
 
 	override initialise(): void {
+		const unknownAction = this.data.actions.UNKNOWN.id
+		this.addEventHook({cause: {type: 'action', action: unknownAction}}, this.triggerUnknownCause)
+		this.addEventHook({action: unknownAction}, this.triggerUnknownCause)
+
+		const unknownStatus = this.data.statuses.UNKNOWN.id
+		this.addEventHook({cause: {type: 'status', status: unknownStatus}}, this.triggerUnknownCause)
+		this.addEventHook({status: unknownStatus}, this.triggerUnknownCause)
+
+		// Listen to all events, but unhook after the first fire.
 		const firstEventHook = this.addEventHook(
 			(event): event is Event => true,
 			event => {
@@ -19,6 +30,14 @@ export class InvalidEvent extends Analyser {
 				this.onFirstEvent(event)
 			}
 		)
+	}
+
+	private triggerUnknownCause() {
+		this.brokenLog.trigger(this, 'unknown action', (
+			<Trans id="core.broken-log.trigger.unknown-action">
+					One or more actions were recorded incorrectly, and could not be parsed.
+			</Trans>
+		))
 	}
 
 	private onFirstEvent(event: Event) {
